ADDING OBJECTS AND SPAWNERS

+++ Pang Remake/Assets/Scripts/AllEvents.cs
@@ -52,6 +52,6 @@ public class QuitButtonClickedEvent : SDD.Events.Event
 #region Score Event
 public class ScoreItemEvent : SDD.Events.Event
 {
-	public float eScore;
+	public int eScore;
 }
 #endregion
